requirements.txt
attrs==19.3.0
chardet==3.0.4
importlib-metadata==0.17
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
selenium==3.141.0
selenium-wire==1.0.11
six==1.12.0
urllib3==1.25.9
wcwidth==0.1.7
zipp==0.5.1
requests